unit db tests updated

+++ .dev/tests/db/class_db_offline_mysql.Test.php
@@ -28,9 +28,11 @@ class class_db_offline_mysql_test extends db_offline_abstract {
 		'name'		=> 'name\'',
 	);
 	public function test_db_prefix() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$this->assertEquals(DB_PREFIX, self::db()->DB_PREFIX);
 	}
 	public function test_fix_table_name() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$this->assertEquals('', self::db()->_fix_table_name(''));
 		$this->assertEquals(self::db()->DB_PREFIX.'sys_admin', self::db()->_fix_table_name('admin'));
 		$this->assertEquals(self::db()->DB_PREFIX.'sys_admin_modules', self::db()->_fix_table_name('admin_modules'));
@@ -51,38 +53,47 @@ class class_db_offline_mysql_test extends db_offline_abstract {
 		$this->assertEquals(self::db()->DB_PREFIX.'dashboards', self::db()->_fix_table_name(self::db()->DB_PREFIX.'dashboards'));
 	}
 	public function test_insert() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->insert('shop_orders', $this->data_safe, $only_sql = true);
 		$this->assertEquals( 'INSERT INTO `t_shop_orders` (`user_id`, `date`, `total_sum`, `name`) VALUES (\'1\', \'1234567890\', \'19,12\', \'name\')', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_insert_safe() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->insert_safe('shop_orders', $this->data_not_safe, $only_sql = true);
 		$this->assertEquals( 'INSERT INTO `t_shop_orders` (`user_id`, `date`, `total_sum`, `name`) VALUES (\'1\', \'1234567890\', \'19,12\', \'name\\\'\')', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_replace() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->replace('shop_orders', $this->data_safe, $only_sql = true);
 		$this->assertEquals( 'REPLACE INTO `t_shop_orders` (`user_id`, `date`, `total_sum`, `name`) VALUES (\'1\', \'1234567890\', \'19,12\', \'name\')', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_replace_safe() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->replace_safe('shop_orders', $this->data_not_safe, $only_sql = true);
 		$this->assertEquals( 'REPLACE INTO `t_shop_orders` (`user_id`, `date`, `total_sum`, `name`) VALUES (\'1\', \'1234567890\', \'19,12\', \'name\\\'\')', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_insert_ignore() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->insert_ignore('shop_orders', $this->data_safe, $only_sql = true);
 		$this->assertEquals( 'INSERT IGNORE INTO `t_shop_orders` (`user_id`, `date`, `total_sum`, `name`) VALUES (\'1\', \'1234567890\', \'19,12\', \'name\')', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_insert_on_duplicate_key_update() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->insert_on_duplicate_key_update('shop_orders', $this->data_safe, $only_sql = true);
 		$this->assertEquals( 'INSERT INTO `t_shop_orders` (`user_id`, `date`, `total_sum`, `name`) VALUES (\'1\', \'1234567890\', \'19,12\', \'name\') ON DUPLICATE KEY UPDATE `user_id` = VALUES(`user_id`), `date` = VALUES(`date`), `total_sum` = VALUES(`total_sum`), `name` = VALUES(`name`)', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_update() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->update('shop_orders', $this->data_safe, 'id=1', $only_sql = true);
 		$this->assertEquals( 'UPDATE `t_shop_orders` SET `user_id` = \'1\', `date` = \'1234567890\', `total_sum` = \'19,12\', `name` = \'name\' WHERE id=1', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_update_safe() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$sql = self::db()->update_safe('shop_orders', $this->data_not_safe, 'id=1', $only_sql = true);
 		$this->assertEquals( 'UPDATE `t_shop_orders` SET `user_id` = \'1\', `date` = \'1234567890\', `total_sum` = \'19,12\', `name` = \'name\\\'\' WHERE id=1', str_replace(PHP_EOL, '', $sql) );
 	}
 	public function test_delete() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$this->assertFalse(self::db()->delete('table', '', $as_sql = true));
 		$this->assertEquals( 'DELETE FROM `t_table` WHERE `id` = \'1\'', self::db()->delete('table', 1, $as_sql = true));
 		$this->assertEquals( 'DELETE FROM `t_table` WHERE `id` = \'1\'', self::db()->delete('table', 'id=1', $as_sql = true));
@@ -92,6 +103,7 @@ class class_db_offline_mysql_test extends db_offline_abstract {
 #		$this->assertEquals( '', self::db()->delete('table', 'id between 1 and 5', $as_sql = true));
 	}
 	public function test_es() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$this->assertEquals( '', self::db()->es(false));
 		$this->assertEquals( '', self::db()->es(''));
 		$this->assertEquals( 'name\\\'', self::db()->es('name\''));
@@ -107,11 +119,13 @@ class class_db_offline_mysql_test extends db_offline_abstract {
 		$this->assertEquals( 'name\\\'', self::db()->real_escape_string('name\''));
 	}
 	public function test_escape_key() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$this->assertEquals( '``', self::db()->escape_key(false));
 		$this->assertEquals( '``', self::db()->escape_key(''));
 		$this->assertEquals( '`name`', self::db()->escape_key('name'));
 	}
 	public function test_escape_val() {
+		if ($this->_need_skip_test(__FUNCTION__)) { return ; }
 		$this->assertEquals( '\'\'', self::db()->escape_val(false));
 		$this->assertEquals( '\'\'', self::db()->escape_val(''));
 		$this->assertEquals( '\'text\'', self::db()->escape_val('text'));
